Sort out webclient tests.
